Projeto-PI-Task-Manager/back-end/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
Werkzeug==3.1.3
